+++ app/src/join/JoinRoute.tsx
@@ -1,0 +1,6 @@
+import { useParams } from 'react-router';
+import { JoinPage } from './JoinPage';
+
+export function JoinRoute() {
+  return <JoinPage code={useParams()['code'] ?? ''} />;
+}